Made a start on magnification map
started magnification map. Marked all pixels with unique RBG values. Now will need to test how many of each appear in the lensed image to get magnification for each pixel and plot that as a colour map.

Good idea would be to then vary rc and eps and see that changes the map.