Study of ( star + big planet transit) lensing with misalignments
Found that misalignments of the star + big planet system w.r.t lensing object cause a noticeable difference in heights of second and third peaks in light curves (ones affected by the planet).
Found the ratios of these heights and returned to user. The ratio decreases as the misalignment increases.

Also, cleaned up the binary system simulation with same methods as I did previously for the planet + star system.

Next, need to

---- further investigate the epsilon parameter study.
---- Investigate the physical importance of ratio of peaks in misaligned lensed binary
---- Investigate misaligned binary with star + star (not black and white), more realistic
---- clean up the codes, collecting all ones that will be used in the end into one final file